feat(Identity): Integrate service into docker-compose
This commit integrates the identity service into the
`docker-compose.yaml` file.  It defines a service named `identity` that
builds from the `./identity` directory and maps port 3000 of the
container to port 3000 on the host.  Additionally it adds a `develop`
section to the `docker-compose.yaml` to handle volume mounting for
development purposes.

The `Dockerfile` for the identity service has been simplified to use
`npm` instead of `pnpm` for the sake of simplicity for demonstration
purposes.  If pnpm is desired, it can be implemented with minor
adjustments.  The `.dockerignore` has also been updated to exclude the
`Dockerfile` from being included in the image build.