feat(identity): Implement account login endpoint
This commit implements the `/account/login` endpoint for the identity
service. This endpoint allows users to log in using their email and
password.  Password verification is performed using Argon2.  If the
credentials are valid, the user's account ID is returned; otherwise, an
empty object is returned.  Drizzle-ORM is used to query the database.