feat: Add reverse proxy with Nginx
This commit introduces a reverse proxy using Nginx to handle routing
requests to the identity service.  A new `reverse-proxy` service is
added to `docker-compose.yaml`, building from a `Dockerfile` that copies
the `nginx.conf` file. The `nginx.conf` file is configured to route
requests to `/account` and `/profile` to the identity service running on
port 3000.  The main port exposed is 80, for HTTP traffic. The
`identity` service no longer exposes a port directly. A `develop`
section is included in `reverse-proxy` for hot reloading the
configuration file.