feat: Add MongoDB database for stores service
This commit adds a MongoDB database container (`stores-db`) to the `docker-compose.yaml` file. This database will be used to store data for the `stores` service (which is assumed to exist, but not yet defined in `docker-compose`). The container uses environment variables to configure the database name and credentials. A healthcheck is implemented to verify the database's availability.  This assumes a future implementation of the `stores` service will use this database.