feat(reverse-proxy): Add stores service routes
This commit adds routes for the `/api/store` and `/api/product`
endpoints to the `nginx.conf` file. These routes will proxy requests to
the corresponding endpoints of the `stores` service, allowing the
reverse proxy to route requests to both the identity and stores
services.  The `reverse-proxy` service now depends on the `stores`
service.